+++ CHANGELOG.md
@@ -15,6 +15,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 # [0.1.x]
 
+- Updated chuck code base to `1.5.4.3-dev (chai)`
+
+- Added `ConvRev` and `Line` chugins
+
 - Updated chuck code base and examples to `1.5.4.2-dev (chai)`
 
 - Updated chuck code base to `1.5.3.2-dev (chai)`